refact : widgets/interview-lobby 불필요한 이미지 리렌더링 제거
