+++ client/src/widgets/interview/ui/interview-lobby/wellcome/index.ts
@@ -1,0 +1,1 @@
+export { default as WellcomeContent } from './WellcomeContent';
